homestay_management/booking_service/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
pydantic==1.8.2
requests==2.26.0
werkzeug==2.0.3
flask-cors